GameGroup class created by extracting functionality from AmoebaTrainer
